+++ backend/canisters/group/impl/src/lifecycle/init.rs
@@ -32,7 +32,6 @@ fn init(args: Args) {
         args.user_index_canister_id,
         args.local_user_index_canister_id,
         args.notifications_canister_id,
-        args.ledger_canister_id,
         args.proposals_bot_user_id,
         args.test_mode,
         args.permissions,