+++ backend/canisters/group_index/impl/src/lifecycle/init.rs
@@ -20,7 +20,6 @@ fn init(args: Args) {
         args.local_group_index_canister_wasm,
         args.user_index_canister_id,
         args.cycles_dispenser_canister_id,
-        args.ledger_canister_id,
         args.proposals_bot_user_id,
         args.test_mode,
     );