+++ backend/canisters/local_group_index/impl/src/updates/c2c_create_group.rs
@@ -82,7 +82,6 @@ fn prepare(args: Args, runtime_state: &mut RuntimeState) -> Result<PrepareOk, Re
         user_index_canister_id: runtime_state.data.user_index_canister_id,
         local_user_index_canister_id: runtime_state.data.local_user_index_canister_id,
         notifications_canister_id: runtime_state.data.notifications_canister_id,
-        ledger_canister_id: runtime_state.data.ledger_canister_id,
         proposals_bot_user_id: runtime_state.data.proposals_bot_user_id,
         avatar: args.avatar,
         test_mode: runtime_state.data.test_mode,