+++ backend/canisters/user/api/src/lifecycle/init.rs
@@ -9,7 +9,6 @@ pub struct Args {
     pub user_index_canister_id: CanisterId,
     pub local_user_index_canister_id: CanisterId,
     pub notifications_canister_id: CanisterId,
-    pub ledger_canister_id: CanisterId,
     pub wasm_version: Version,
     pub username: String,
     pub test_mode: bool,