+++ backend/canisters/user/impl/src/lifecycle/init.rs
@@ -19,7 +19,6 @@ fn init(args: Args) {
         args.local_user_index_canister_id,
         args.group_index_canister_id,
         args.notifications_canister_id,
-        args.ledger_canister_id,
         args.username,
         args.test_mode,
         env.now(),