+++ backend/canisters/user_index/api/src/lifecycle/init.rs
@@ -13,7 +13,6 @@ pub struct Args {
     pub notifications_index_canister_id: CanisterId,
     pub cycles_dispenser_canister_id: CanisterId,
     pub open_storage_index_canister_id: CanisterId,
-    pub ledger_canister_id: CanisterId,
     pub proposals_bot_user_id: UserId,
     pub local_group_index_canister_ids: Vec<CanisterId>,
     pub wasm_version: Version,