+++ backend/canisters/group_index/impl/src/updates/add_local_group_index_canister.rs
@@ -48,7 +48,6 @@ fn prepare(args: &Args, runtime_state: &RuntimeState) -> Result<PrepareResult, R
                 group_index_canister_id: runtime_state.env.canister_id(),
                 notifications_canister_id: args.notifications_canister_id,
                 cycles_dispenser_canister_id: runtime_state.data.cycles_dispenser_canister_id,
-                ledger_canister_id: runtime_state.data.ledger_canister_id,
                 proposals_bot_user_id: runtime_state.data.proposals_bot_user_id,
                 test_mode: runtime_state.data.test_mode,
             },