+++ backend/canisters/local_user_index/impl/src/lib.rs
@@ -83,7 +83,6 @@ impl RuntimeState {
                 group_index: self.data.group_index_canister_id,
                 notifications: self.data.notifications_canister_id,
                 cycles_dispenser: self.data.cycles_dispenser_canister_id,
-                icp_ledger: self.data.ledger_canister_id,
             },
         }
     }
@@ -99,7 +98,6 @@ struct Data {
     pub notifications_canister_id: CanisterId,
     pub cycles_dispenser_canister_id: CanisterId,
     pub canisters_requiring_upgrade: CanistersRequiringUpgrade,
-    pub ledger_canister_id: CanisterId,
     pub canister_pool: canister::Pool,
     pub total_cycles_spent_on_canisters: Cycles,
     pub user_event_sync_queue: CanisterEventSyncQueue<UserEvent>,
@@ -122,7 +120,6 @@ impl Data {
         group_index_canister_id: CanisterId,
         notifications_canister_id: CanisterId,
         cycles_dispenser_canister_id: CanisterId,
-        ledger_canister_id: CanisterId,
         canister_pool_target_size: u16,
         test_mode: bool,
     ) -> Self {
@@ -134,7 +131,6 @@ impl Data {
             group_index_canister_id,
             notifications_canister_id,
             cycles_dispenser_canister_id,
-            ledger_canister_id,
             canisters_requiring_upgrade: CanistersRequiringUpgrade::default(),
             canister_pool: canister::Pool::new(canister_pool_target_size),
             total_cycles_spent_on_canisters: 0,
@@ -173,5 +169,4 @@ pub struct CanisterIds {
     pub group_index: CanisterId,
     pub notifications: CanisterId,
     pub cycles_dispenser: CanisterId,
-    pub icp_ledger: CanisterId,
 }