+++ backend/canisters/local_user_index/impl/src/lifecycle/init.rs
@@ -23,7 +23,6 @@ fn init(args: Args) {
         args.group_index_canister_id,
         args.notifications_canister_id,
         args.cycles_dispenser_canister_id,
-        args.ledger_canister_id,
         canister_pool_target_size,
         args.test_mode,
     );