+++ backend/canisters/local_group_index/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.602](https://github.com/open-ic/open-chat/releases/tag/v2.0.602-local_group_index)] - 2023-02-17
+
 ### Added
 
 - Added `c2c_reinstall_group` for recovering groups which are un-upgradable due to bug in `pre_upgrade` ([#3128](https://github.com/open-ic/open-chat/pull/3128))